Extract spec visualization widgets and rename render package
- Move simapp-free Streamlit widget rendering into
  spec_visualization/visualization_widgets.py as render_visualization_widgets.
- Rename review_gui/render/ to spec_visualization/ and update imports.
- Keep render_visualization_panel as the simapp-coupled entry point.
